requirements.txt
openai>=1.91.0
python-dotenv
watchdog
pillow
